Refactor the client secret integration tests for robustness and reuse
