+++ modules/integration/tests-integration/tests-backend/src/test/java/org/wso2/identity/integration/test/rest/api/server/application/management/v1/ApplicationManagementClientSecretSuccessTest.java
@@ -24,7 +24,6 @@ import io.restassured.response.Response;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.http.HttpHeaders;
 import org.apache.http.HttpStatus;
-import org.json.JSONArray;
 import org.json.JSONObject;
 import org.testng.Assert;
 import org.testng.annotations.AfterClass;
@@ -34,7 +33,6 @@ import org.wso2.carbon.automation.engine.context.TestUserMode;
 
 import java.time.Instant;
 import java.util.Collections;
-import java.util.Optional;
 
 import static org.hamcrest.CoreMatchers.equalTo;
 import static org.hamcrest.Matchers.notNullValue;
@@ -464,46 +462,6 @@ public class ApplicationManagementClientSecretSuccessTest extends ApplicationMan
                 .assertThat()
                 .statusCode(HttpStatus.SC_NO_CONTENT);
         createdAppId = null;
-    }
-
-    /**
-     * Build an OAuth application creation payload carrying the given client secret expiry.
-     *
-     * @param appName               Name of the application.
-     * @param clientSecretExpiresAt Client secret expiry as Unix epoch seconds; null to omit the field.
-     * @return Application creation payload.
-     * @throws Exception If an error occurs while reading or building the payload.
-     */
-    private String buildOAuthAppPayload(String appName, Long clientSecretExpiresAt) throws Exception {
-
-        JSONObject payload = new JSONObject(readResource("create-oauth-app.json"));
-        payload.put("name", appName);
-        if (clientSecretExpiresAt != null) {
-            payload.getJSONObject("inboundProtocolConfiguration").getJSONObject("oidc")
-                    .put("clientSecretExpiresAt", clientSecretExpiresAt);
-        }
-        return payload.toString();
-    }
-
-    /**
-     * Build an OIDC inbound update payload from the current configuration of the application.
-     *
-     * @param responseOfGet Response of the OIDC inbound configuration retrieval.
-     * @return OIDC inbound update payload.
-     * @throws Exception If an error occurs while building the payload.
-     */
-    private JSONObject buildOIDCUpdatePayload(Response responseOfGet) throws Exception {
-
-        JsonPath currentConfig = responseOfGet.jsonPath();
-        JSONObject payload = new JSONObject();
-        payload.put("clientId", currentConfig.getString("clientId"));
-        payload.put("clientSecret", currentConfig.getString("clientSecret"));
-        payload.put("grantTypes", new JSONArray(currentConfig.getList("grantTypes")));
-        payload.put("callbackURLs", new JSONArray(currentConfig.getList("callbackURLs")));
-        payload.put("publicClient", currentConfig.getBoolean("publicClient"));
-        payload.put("allowedOrigins", new JSONArray(Optional.ofNullable(currentConfig.getList("allowedOrigins"))
-                .orElse(Collections.emptyList())));
-        return payload;
     }
 
     private String getOIDCConfigPath() {
