--- modules/integration/tests-integration/tests-backend/src/test/resources/org/wso2/identity/integration/test/oauth2/multi-level-org-management-app-authorized-apis.json
@@ -1,40 +1,0 @@
-{
-  "/api/server/v1/organizations": [
-    "internal_organization_view",
-    "internal_organization_create",
-    "internal_organization_update",
-    "internal_organization_delete"
-  ],
-  "/o/api/server/v1/organizations": [
-    "internal_org_organization_view",
-    "internal_org_organization_create",
-    "internal_org_organization_delete"
-  ],
-  "/o/api/server/v1/applications": [
-    "internal_org_application_mgt_update",
-    "internal_org_application_mgt_view",
-    "internal_org_application_mgt_create",
-    "internal_org_application_mgt_delete",
-    "internal_org_application_internal_api_update",
-    "internal_org_application_business_api_update"
-  ],
-  "/o/api/server/v1/api-resources": [
-    "internal_org_api_resource_view"
-  ],
-  "/o/scim2/Users": [
-    "internal_org_user_mgt_view",
-    "internal_org_user_mgt_list",
-    "internal_org_user_mgt_create",
-    "internal_org_user_mgt_update",
-    "internal_org_user_mgt_delete"
-  ],
-  "/o/scim2/Roles": [
-    "internal_org_role_mgt_view",
-    "internal_org_role_mgt_create",
-    "internal_org_role_mgt_update",
-    "internal_org_role_mgt_delete",
-    "internal_org_role_mgt_permissions_update",
-    "internal_org_role_mgt_users_update",
-    "internal_org_role_mgt_groups_update"
-  ]
-}
